STY: Refactor all Term functionality into separate file.
